+++ src/ResourceManager/TrafficManager/Commands.TrafficManager2.Test/ScenarioTests/Common.ps1
@@ -94,3 +94,14 @@ function TestSetup-AddEndpoint($endpointName, $profile)
 
 	return $profile
 }
+
+<#
+.SYNOPSIS
+Cleans the created resource groups
+#>
+function TestCleanup-RemoveResourceGroup($rgname)
+{
+    if ([Microsoft.Azure.Test.HttpRecorder.HttpMockServer]::Mode -ne [Microsoft.Azure.Test.HttpRecorder.HttpRecorderMode]::Playback) {
+        Remove-AzureRmResourceGroup -Name $rgname -Force
+    }
+}